requirements.txt
requests>=2.28.0
colorama>=0.4.6
google-genai>=1.0.0
python-dotenv>=1.0.0
